+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Improved
 - created API library usage for all API functions
 
+### Updated
+- updated Omniva API library to v1.0.11
+
 ## [1.14.2]
 ### Fixed
 - fixed error when Order create manualy from admin area